requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
autopep8==1.6.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
colorama==0.4.4
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
datetime-truncate==1.1.1
defusedxml==0.7.1
Django==4.1.1
django-3-jet==1.0.8
django-4-jet==1.0.9
django-admin-charts==1.3.0
django-allauth==0.51.0
django-bower==5.2.0
django-ckeditor==6.5.0
django-colorfield==0.8.0
django-extensions==3.1.3
django-jazzmin==2.4.8
django-jet==1.0.8
django-js-asset==2.0.0
django-memoize==2.3.1
django-mptt==0.13.4
django-multiselectfield==0.1.12
django-ninja==0.19.1
django-nvd3==0.9.7
django-rest-framework==0.1.0
django-silk==4.2.0
django-simple-history==3.0.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
feedparser==6.0.10
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
google-api-python-client==1.4.1
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gprof2dot==2021.2.21
grpcio==1.51.1
httplib2==0.20.4
idna==3.3
install==1.3.4
Jinja2==3.0.2
keras==2.11.0
Keras-Applications==1.0.8
keras-vggface==0.6
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.6.2
mtcnn==0.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.1
opencv-python==4.7.0.68
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.2.0
pretty-errors==1.2.24
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.4
pydot==1.4.2
pyExcelerator==0.6.4.1
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.4
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-jose==3.3.0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.7.2
scipy==1.9.3
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
termcolor==2.1.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.2
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.14.1
